Refactor grpc service tests
Starting an actual server instance in a separate thread and unit
testing service logic seems too heavy an approach. Also, the tests
are not deterministic due to server unavailability and other race
conditions. Since this is against the nature of unit tests, the
service class is directly tested.

However, there is a bug in gRPC that prevents the ServerContext
from being instantiated outside of a properly started-up server.
To fix that, a forked version of gRPC is used as a dependency for
now.