Improve API design for better usability (#3)
* GetDetectableObjects:
Remove needless parameter that might confuse users. Keep this rpc
simple and to the point. Language of docs, tests and names cleaned
up so that its easier to understand.

* DetectImage:
Removed object_to_detect because it is not useful for now and it
only gets in the way of users. Cleaned up the language to make 
service more straightforward and understandable.

* Updated docs